move savefile to new io/Dialog file
