+++ src/renderer/io/Data.ts
@@ -1,20 +1,19 @@
 import { Capacitor } from '@capacitor/core'
 import path from 'path'
+import { promisify } from 'util'
 
 const PLATFORM = Capacitor.platform
 const platformNotSupportedMessage = `ERROR - PLATFORM NOT SUPPORTED: "${PLATFORM}" `
 
 // variables used by electron
 let fs: typeof import('fs')
-let promisify: typeof import('util').promisify
+
 let electron: typeof import('electron')
 
 let userDataPath: string
 
 if (PLATFORM == 'electron') {
   fs = require('fs')
-  promisify = require('util').promisify
-
   electron = require('electron')
   userDataPath = path.join((electron.app || electron.remote.app).getPath('userData'), 'data')
 }
@@ -43,33 +42,6 @@ const readFile = async function(name: string): Promise<string> {
   }
 }
 
-const saveFile = function(filename, data) {
-  switch (PLATFORM) {
-    case 'web':
-      var blob = new Blob([data], { type: 'text/csv' })
-      if (window.navigator.msSaveOrOpenBlob) {
-        window.navigator.msSaveBlob(blob, filename)
-      } else {
-        var elem = window.document.createElement('a')
-        elem.href = window.URL.createObjectURL(blob)
-        elem.download = filename
-        document.body.appendChild(elem)
-        elem.click()
-        document.body.removeChild(elem)
-      }
-      return
-    case 'electron':
-      const { dialog } = electron.remote
-      dialog.showSaveDialog({
-        defaultPath: filename,
-        buttonLabel: 'Save Pilot',
-      })
-      return
-    default:
-      throw new Error(platformNotSupportedMessage)
-  }
-}
-
 const exists = async function(name: string): Promise<boolean> {
   switch (PLATFORM) {
     case 'web':
@@ -87,7 +59,17 @@ const saveData = async function<T>(fileName: string, data: T) {
 
 const loadData = async function<T>(fileName: string) {
   const dataText = await readFile(fileName)
-  return JSON.parse(dataText) as T
+  return (JSON.parse(dataText) || []) as T[]
+}
+
+const importData = function<T>(file: File): Promise<T> {
+  return new Promise(resolve => {
+    var fr = new FileReader()
+    fr.onload = e => {
+      resolve(JSON.parse(e.target.result as string) as T)
+    } // CHANGE to whatever function you want which would eventually call resolve
+    fr.readAsText(file)
+  })
 }
 
 const staticPath = function(pathEnd: string) {
@@ -112,4 +94,4 @@ const dataPathMap = {
 
 const USER_DATA_PATH = dataPathMap[PLATFORM]
 
-export { writeFile, readFile, saveFile, saveData, loadData, exists, staticPath, USER_DATA_PATH }
+export { writeFile, readFile, saveData, loadData, importData, exists, staticPath, USER_DATA_PATH }
